🧹 Refactor: Extract shared Taylor series utilities
- Extracted `factorial`, `sine_taylor`, and `cosine_taylor` from various files in `Math/Geometry/Trigonometry/` to `Math/Geometry/Trigonometry/utils/taylor_series.py`.
- Refactored `tan.py`, `cosecant.py`, `secant.py`, and `cosine_rule.py` to use these shared utilities, reducing code duplication.
- Minor cleanup of an unused variable calculation (`n - i - j`) in `trinomial_theorem.py`.
- Formatted and passed all automated tests successfully.
