+++ Tests/test_DiscreteMath.py
@@ -25,17 +25,6 @@ if combinatorics_dir not in sys.path:
 
 from Math.Discrete_Math.Combinatorics.binomial_theorem_general_term import binomial_general_term
 from Math.Discrete_Math.Combinatorics.pascals_triangle import print_pascals_triangle, generate_pascals_triangle
-from Math.Discrete_Math.Combinatorics.binomial_theorem import (
-    expand_binomial,
-    binomial_coefficient,
-)
-from Math.Discrete_Math.Combinatorics.binomial_theorem_general_term import (
-    binomial_general_term,
-)
-from Math.Discrete_Math.Combinatorics.pascals_triangle import (
-    print_pascals_triangle,
-    generate_pascals_triangle,
-)
 from Math.Discrete_Math.Combinatorics.permutation import factorial, n_permute_r
 from Math.Discrete_Math.Combinatorics.trinomial_theorem import (
     expand_trinomial,
@@ -44,24 +33,11 @@ from Math.Discrete_Math.Combinatorics.trinomial_theorem import (
 from Math.Discrete_Math.Combinatorics.trinomial_theorem_general_term import (
     trinomial_general_term,
 )
-from Math.Discrete_Math.Number_Theory.lcm import compute_lcm, prime_factorization_simple
-from Math.Discrete_Math.Number_Theory.gcd import (
-    compute_gcd,
-    prime_factorization_for_gcd,
-)
 from Math.Discrete_Math.Number_Theory.prime_factorisation import prime_factorization
 from Math.Discrete_Math.Number_Theory.partitions_approximation import (
     partition_approximation,
 )
 from Math.Discrete_Math.Number_Theory.partitions import partition
-from Math.Discrete_Math.Combinatorics.binomial_theorem_general_term import binomial_general_term  # noqa: E402
-from Math.Discrete_Math.Combinatorics.pascals_triangle import print_pascals_triangle, generate_pascals_triangle  # noqa: E402
-from Math.Discrete_Math.Combinatorics.permutation import factorial, n_permute_r  # noqa: E402
-from Math.Discrete_Math.Combinatorics.trinomial_theorem import expand_trinomial, trinomial_coefficient  # noqa: E402
-from Math.Discrete_Math.Combinatorics.trinomial_theorem_general_term import trinomial_general_term  # noqa: E402
-from Math.Discrete_Math.Number_Theory.prime_factorisation import prime_factorization  # noqa: E402
-from Math.Discrete_Math.Number_Theory.partitions_approximation import partition_approximation  # noqa: E402
-from Math.Discrete_Math.Number_Theory.partitions import partition  # noqa: E402
 
 
 def test_factorial_zero():
