+++ Tests/test_arctan_security.py
@@ -1,7 +1,5 @@
 import unittest
 from decimal import Decimal
-import decimal
-import math
 
 from Math.Geometry.Trigonometry.Arc_Functions.arctan import calculate_arctan
